refactor(landing): remove dead code, restructure components, standardize exports (#310)
* refactor(landing): remove dead code, restructure components, standardize exports

Behavior-preserving cleanup of the landing app.

- Remove unused PixelBlast (tsx/shaders/css) and FeatureVisualSessionPersistence
- Drop now-unused deps: ogl, postprocessing, three, framer-motion, @types/three
- Restructure src/components into layout/, sections/, ui/ (feature-visuals unchanged)
- Standardize all components on named exports
- Rename src/pages -> src/routes; move routes.tsx -> routes/index.tsx

Verified: lint, tsc -b, bun test (15 pass), build (3 routes prerendered),
verify:prerender all pass. No visual or behavioral change.

Closes #9

* refactor(landing): consolidate @/ shadcn folder into src/

Move vendored shadcn components into src/ and point the @ alias at src/,
matching the standard shadcn convention (@ = src).

- Move @/components/ui/{avatar,infinite-slider}.tsx -> src/components/ui/
- Remove @/lib/utils.ts re-export shim (src/lib/utils.ts is the source)
- Delete the @/ directory
- Repoint @ alias to ./src in vite.config.ts and tsconfig.app.json;
  simplify tsconfig include to ["src"]

@/ imports resolve unchanged (now to src/), so no import edits needed.
shadcn CLI (components.json) keeps working.

Verified: lint, tsc -b, bun test (15 pass), build (3 routes prerendered),
verify:prerender all pass. No behavior change.
