--- landing/@/lib/utils.ts
@@ -1,1 +1,0 @@
-export { cn } from '../../src/lib/utils';
