+++ landing/package.json
@@ -23,19 +23,15 @@
     "@radix-ui/react-avatar": "^1.1.11",
     "@unhead/react": "^3.1.1",
     "clsx": "^2.1.1",
-    "framer-motion": "^12.38.0",
     "hono": "^4.12.23",
     "lucide-react": "^1.17.0",
     "motion": "^12.40.0",
-    "ogl": "^1.0.11",
     "overlayscrollbars-react": "^0.5.6",
-    "postprocessing": "^6.39.1",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
     "react-router": "^7.16.0",
     "react-use-measure": "^2.1.7",
     "tailwind-merge": "^3.6.0",
-    "three": "^0.184.0",
     "vite-plugin-react-ssg": "^0.2.0"
   },
   "devDependencies": {
@@ -46,7 +42,6 @@
     "@types/node": "^24.12.3",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
-    "@types/three": "^0.184.1",
     "@vitejs/plugin-react": "^6.0.1",
     "autoprefixer": "^10.5.0",
     "eslint": "^10.3.0",
