--- landing/src/components/PixelBlast.css
@@ -1,6 +1,0 @@
-.pixel-blast-container {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  overflow: hidden;
-}
