+++ landing/src/components/ui/SectionHeader.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react';
 
-import { cn } from '../lib/utils';
+import { cn } from '../../lib/utils';
 
 type SectionHeaderProps = {
   eyebrow?: ReactNode;
