+++ landing/src/routes/HomePage.tsx
@@ -1,0 +1,17 @@
+import { FeatureSection } from '../components/sections/FeatureSection';
+import { MoreFeaturesSection } from '../components/sections/MoreFeaturesSection';
+import { Hero } from '../components/sections/Hero';
+import { TestimonialsSection } from '../components/sections/TestimonialsSection';
+
+export function HomePage() {
+  return (
+    <>
+      <main id="main-content">
+        <Hero />
+        <FeatureSection />
+        <MoreFeaturesSection />
+        <TestimonialsSection />
+      </main>
+    </>
+  );
+}
