+++ landing/vite.config.ts
@@ -11,7 +11,7 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url))
 export default defineConfig({
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './@'),
+      '@': path.resolve(__dirname, './src'),
     },
   },
   plugins: [
